Somsiad zmienia wagi floorów? Ja tam wole Dijkstre.

+++ CityEvacuation/src/sample/structure/logic/StaticPoint.java
@@ -8,7 +8,7 @@ import java.util.List;
 public abstract class StaticPoint {
     private int x;
     private int y;
-    public int weight;
+    public float weight;
 
     protected Color tileColor;
     protected List<ActionType> actionTypeList;
@@ -46,7 +46,7 @@ public abstract class StaticPoint {
     public abstract List<ActionType> getPossibleActions();
     public abstract boolean interact();
 
-    public StaticPoint(int x, int y, int weight) {
+    public StaticPoint(int x, int y, float weight) {
         this.setX(x);
         this.setY(y);
         this.tileColor();
